Create a base list filter bar component class.
