+++ src/components/gear/gear-list-filter-bar/gear-list-filter-bar.ts
@@ -1,38 +1,21 @@
 import html from "./gear-list-filter-bar.html";
-import { BaseComponent } from "../../base.component";
 import { GearFilterChangedEvent } from "../../../lib/events/gear-filter-changed-event";
 import { EventBus } from "../../../lib/events/event-bus";
 import { GearItem } from "../../../features/gear/gear-item";
 import { GearCategoryChangedEvent } from "../../../lib/events/gear-category-changed-event";
 import { AppEvent } from "../../../lib/events/app-event";
 import { AppEventListener } from "../../../lib/events/app-event-listener-interface";
-import { IUnitOfWork } from "../../../lib/common/data-access/unit-of-work-interface";
-import { appInjector } from "../../../lib/infrastructure/app-injector";
-import { UnitOfWork } from "../../../lib/data-access/unit-of-work";
-import { Source } from "../../../features/sources/source";
-import { CustomSelectComponent } from "../../custom-select/custom-select";
-import { SelectOption } from "../../../lib/selects/select-option";
+import { BaseListFilterBarComponent } from "../../base-list-filter-bar/base-list-filter-bar";
 
-export class GearListFilterBarComponent extends BaseComponent {
+export class GearListFilterBarComponent extends BaseListFilterBarComponent {
     private readonly activeButtonCssClass = "active-filter-button";
     private readonly inactiveButtonCssClass = "inactive-filter-button";
 
-    private unitOfWork: IUnitOfWork;
-
     private activeCategory: string;
-    private currentSearch: string;
-    private currentSourceId: number;
-
-    public get sourcesSelectElement(): CustomSelectComponent {
-        return this.shadow.querySelector("#sourcesFilter") as CustomSelectComponent;
-    }
 
     constructor() {
         super();
-        this.unitOfWork = appInjector.injectClass(UnitOfWork);
         this.activeCategory = GearItem.gearCategory;
-        this.currentSearch = "";
-        this.currentSourceId = 0;
     }
 
     public connectedCallback() {
@@ -46,36 +29,10 @@ export class GearListFilterBarComponent extends BaseComponent {
         EventBus.instance.unregister(GearCategoryChangedEvent.name, this.handleGearCategoryChangedEvent);
     }
 
-    private configureSourcesFilter() {
-        this.sourcesSelectElement.onOptionChange = this.handleOnSourcesSelectChanged;
-        this.sourcesSelectElement.containerCssClassList.add("h-full");
-        this.sourcesSelectElement.customSelectButtonCssClassList.add("h-full");
-        this.populateSourcesFilter();
-    }
-
-    public handleOnSourcesSelectChanged = (newValue: SelectOption) => {
-        const selectedValue = newValue.value;
-
-        this.currentSourceId = parseInt(selectedValue);
-
-        this.dispatchGearFilterChangedEvent();
-    };
-
-    private dispatchGearFilterChangedEvent() {
+    protected dispatchFilterChangedEvent() {
         const appEvent = new GearFilterChangedEvent(this.activeCategory, this.currentSearch, this.currentSourceId);
 
         EventBus.instance.dispatch(appEvent);
-    }
-
-    private populateSourcesFilter() {
-        const sources = this.unitOfWork.repo(Source).list();
-        const sourceOptions = [new SelectOption("0", "All")];
-
-        for (let source of sources) {
-            sourceOptions.push(new SelectOption(source.id.toString(), source.name));
-        }
-
-        this.sourcesSelectElement.populateOptions(sourceOptions);
     }
 
     private handleGearCategoryChangedEvent: AppEventListener = (event: AppEvent) => {
@@ -83,18 +40,8 @@ export class GearListFilterBarComponent extends BaseComponent {
 
         this.activeCategory = gearCategoryChangedEvent.category;
 
-        this.dispatchGearFilterChangedEvent();
+        this.dispatchFilterChangedEvent();
     };
-
-    public handleOnSearchBoxKeyUp(event: KeyboardEvent) {
-        // Ignore shift key up events, otherwise two GearFilterChangedEvents are triggered when
-        // typing an UPPERCASE character into the search box.
-        if (event.shiftKey === true) return;
-
-        this.currentSearch = (event.target as HTMLInputElement).value;
-
-        this.dispatchGearFilterChangedEvent();
-    }
 }
 
 customElements.define("gear-list-filter-bar", GearListFilterBarComponent);
