+++ src/components/npcs/npc-list-filter-bar/npc-list-filter-bar.html
@@ -1,8 +1,10 @@
-<div class="flex">
+<div class="grid grid-cols-3 gap-2 my-2">
     <input
         id="searchBox"
         placeholder="Search for NPC by name . . ."
-        class="grow border-3 border-black my-2 px-2 py-1 hover:bg-gray-200 transition duration-150 ease-in-out focus:outline-0"
-        onkeyup="this.onSearchBoxKeyUp(event)"
+        class="col-span-2 border-3 border-black px-2 py-1 hover:bg-gray-200 transition duration-150 ease-in-out focus:outline-0"
+        onkeyup="this.handleOnSearchBoxKeyUp(event)"
     />
+
+    <custom-select id="sourcesFilter"></custom-select>
 </div>
